collapsed Subject into GameController
