+++ app/src/main/java/team5/explodingkittens/controller/Observer.java
@@ -4,7 +4,7 @@ import team5.explodingkittens.controller.notification.Notification;
 
 /**
  * An interface for each type of object that can observe
- * a {@link Subject}, and receive notifications from it.
+ * a {@link GameController}, and receive notifications from it.
  *
  * @author Duncan McKee, Maura Coriale
  */
